backend/requirements.txt
Django>=3.2
djangorestframework
djoser
drf-extra-fields
Pillow